refactor: moved AverageStars to Show dir
