+++ resources/js/Pages/Apartment/Show/AverageStars.tsx
