+++ resources/js/Pages/Apartment/Show/Review.tsx
@@ -1,6 +1,6 @@
 import { PageProps, Review as ReviewType } from "@/types";
 import { Card, CardContent, CardHeader } from "@/components/ui/card";
-import AverageStars from "@/components/AverageStars";
+import AverageStars from "@/Pages/Apartment/Show/AverageStars";
 import UserAvatar from "@/components/UserAvatar";
 import { formatDistanceToNow } from "date-fns";
 import { Link, usePage } from "@inertiajs/react";
